+++ CHANGELOG.rst
@@ -1,6 +1,16 @@
 Changelog for Nautilai
 ======================
 
+
+0.1.43 (unreleased)
+-------------------
+
+Changed:
+^^^^^^^^
+- Log files are now located in C:\Users\<user>\Documents\Nautilai\Logs
+- Platemap preview supports all 24 and 96 well plates
+
+
 0.1.42 (2024-04-10)
 -------------------